Cleaned up my warnings and my code, added a few asserts.

+++ src/engine/inc/Log.h
@@ -36,7 +36,7 @@ public:
 
 		if ((pieceByte & 128) == 0)
 		{
-			piece = toupper(piece);
+			piece = (byte)toupper(piece);
 		}
 
 		return piece;
